Add files via upload
Testing if it works now, this should be good except two things on lines Line 605, 613, 630, 638, need to figure out how to print the 3 weapon numbers in errorCheck function. Two, need to get the name of the murderer, the weapon, and the room and put them into a string for displaying during their dialogue lines in the fight.